implement comments for PR
