+++ wremnants/histmaker_tools.py
@@ -197,3 +197,28 @@ def write_analysis_output(results, outfile, args):
     logger.info(f"Output saved in {outfile}")
 
     return outfile
+
+
+def get_run_lumi_edges(nRunBins, era):
+    if era == "2016PostVFP":
+        if nRunBins == 2:
+            run_edges = [278768, 280385, 284044]
+            lumi_edges = [0.0, 0.48013, 1.0]
+        elif nRunBins == 3:
+            run_edges = [278768, 279767, 283270, 284044]
+            lumi_edges = [0.0, 0.25749, 0.72954, 1.0]
+        elif nRunBins == 4:
+            run_edges = [278768, 279767, 280385, 283270, 284044]
+            lumi_edges = [0.0, 0.25749, 0.48013, 0.72954, 1.0]
+        elif nRunBins == 5:
+            run_edges = [278768, 279588, 280017, 282037, 283478, 284044]
+            lumi_edges = [0.0, 0.13871, 0.371579, 0.6038544, 0.836724, 1.0]
+        else:
+            raise NotImplementedError(
+                f"Invalid number of bins ({nRunBins}) passed to --nRunBins."
+            )
+    else:
+        raise NotImplementedError(
+            f"Function get_run_lumi_edges() does not yet support era {era}."
+        )
+    return run_edges, lumi_edges
