+++ wremnants/syst_tools.py
@@ -1,7 +1,9 @@
 import collections.abc
+import pickle
 import re
 
 import hist
+import lz4.frame
 import numpy as np
 import ROOT
 
@@ -2620,3 +2622,44 @@ def add_helicity_hists(
         results.append(gen_theoryAgnostic)
 
     return df
+
+
+def scale_hist_up_down(h, scale):
+    hUp = hh.scaleHist(h, scale)
+    hDown = hh.scaleHist(h, 1 / scale)
+
+    hVar = hist.Hist(
+        *[a for a in h.axes],
+        common.down_up_axis,
+        storage=hist.storage.Weight(),
+    )
+    hVar.values(flow=True)[...] = np.stack(
+        [hDown.values(flow=True), hUp.values(flow=True)], axis=-1
+    )
+    hVar.variances(flow=True)[...] = np.stack(
+        [hDown.variances(flow=True), hUp.variances(flow=True)], axis=-1
+    )
+    return hVar
+
+
+def scale_hist_up_down_corr_from_file(h, corr_file=None, corr_hist=None):
+    # FIXME: this might not be thread safe, but it is a test for now
+    with lz4.frame.open(corr_file) as f:
+        corrs = pickle.load(f)
+    boost_corr = corrs[corr_hist]
+
+    hUp = hh.multiplyHists(h, boost_corr)
+    hDown = hh.divideHists(h, boost_corr)
+
+    hVar = hist.Hist(
+        *[a for a in h.axes],
+        common.down_up_axis,
+        storage=hist.storage.Weight(),
+    )
+    hVar.values(flow=True)[...] = np.stack(
+        [hDown.values(flow=True), hUp.values(flow=True)], axis=-1
+    )
+    hVar.variances(flow=True)[...] = np.stack(
+        [hDown.variances(flow=True), hUp.variances(flow=True)], axis=-1
+    )
+    return hVar
